+++ src/compiler/emitter/mod.rs
@@ -2,5 +2,7 @@ mod lines_0_1000;
 pub use lines_0_1000::*;
 mod lines_1000_1500;
 pub use lines_1000_1500::*;
-mod lines_1500_end;
-pub use lines_1500_end::*;
+mod lines_1500_2000;
+pub use lines_1500_2000::*;
+mod lines_2000_end;
+pub use lines_2000_end::*;
